fix: move remoteclient api modules to separate module
